requirements.txt
django~=4.2.17
django-environ~=0.11.2
Pillow==9.0.*
django-admin-sortable2~=2.2.4